Major Refactor
Used the sveltekit library template for building and moving away from web components for the screen and editor. Sources will stay as web components though for ease of integration.
