--- package/README.md
@@ -1,3 +1,0 @@
-# @sparklapse/dots
-
-The framework for making creator tools.
