--- package/lib/overlay/scene/types.d.ts
@@ -1,1 +1,0 @@
-export { Sources, Transform } from ".";
