--- package/lib/overlay/screen/index.ts
@@ -1,3 +1,0 @@
-import { DotsScreen } from "./web";
-
-customElements.define("dots-screen", DotsScreen);
