+++ package/src/lib/overlay/editor/Inspector.svelte
