--- package/src/main.ts
@@ -1,7 +1,0 @@
-import "./global.css";
-
-import App from "./App.svelte";
-
-new App({
-  target: document.getElementById("app")!,
-});
