+++ package/src/routes/+layout.svelte
@@ -1,0 +1,7 @@
+<script>
+  import "./global.pcss";
+</script>
+
+<div class="flex flex-col h-screen overflow-x-hidden">
+  <slot />
+</div>
