+++ package/src/routes/+page.svelte
@@ -1,0 +1,1 @@
+<h1>@sparklapse/dots</h1>
