+++ package/tailwind.config.js
@@ -1,8 +1,8 @@
-import { dotsUI } from "./lib/ui";
+import { dotsUI } from "./src/lib/ui";
 
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ["./index.html", "./{src,lib}/**/*.{js,ts,svelte}"],
+  content: ["./index.html", "./src/**/*.{js,ts,svelte}"],
   theme: {
     extend: {},
   },
