+++ package/tests/test.ts
@@ -1,0 +1,6 @@
+import { expect, test } from "@playwright/test";
+
+test("index page has expected h1", async ({ page }) => {
+  await page.goto("/");
+  await expect(page.getByRole("heading", { name: "Welcome to SvelteKit" })).toBeVisible();
+});
