+++ package/vite.config.ts
@@ -1,26 +1,9 @@
-import { defineConfig } from "vite";
-import { svelte } from "@sveltejs/vite-plugin-svelte";
-import { resolve } from "node:path";
+import { sveltekit } from "@sveltejs/kit/vite";
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  resolve: {
-    alias: {
-      $lib: resolve("./lib"),
-      $dist: resolve("./dist"),
-    },
+  plugins: [sveltekit()],
+  test: {
+    include: ["src/**/*.{test,spec}.{js,ts}"],
   },
-  plugins: [
-    svelte({
-      compilerOptions: {
-        customElement: true,
-      },
-    }),
-    {
-      // Since web components can't be redefined, full reload
-      name: "full-reload",
-      handleHotUpdate: ({ server }) => {
-        server.ws.send({ type: "full-reload" });
-      },
-    },
-  ],
 });
